Arreglado bug de eliminarPartidas a proyectos y acciones que no permitía eliminar un PresupuestoPartidas Arreglado varios agujeros de seguridad en las funciones de eliminación. Cambiada etiqueta del codigo ncm.

+++ compras/protected/models/PresupuestoPartidaAcciones.php
@@ -67,8 +67,9 @@ class PresupuestoPartidaAcciones extends CActiveRecord
 
 	// Delete cascade / Borrado en cascada
 	public function afterDelete(){
-	    foreach($this->presupuestoPartidas as $c)
-	        $c->delete();
+		if($this->scenario != 'cascadaPartida')
+	    	foreach($this->presupuestoPartidas as $c)
+	        	$c->delete();
 	    return parent::afterDelete();
 	}
