+++ compras/protected/models/PresupuestoPartidaProyecto.php
@@ -47,8 +47,9 @@ class PresupuestoPartidaProyecto extends CActiveRecord
 
 	// Delete cascade / Borrado en cascada
 	public function afterDelete(){
-	    foreach($this->presupuestoPartida as $c)
-	        $c->delete();
+		if($this->scenario != 'cascadaPartida')
+	   		foreach($this->presupuestoPartida as $c)
+	        	$c->delete();
 	    return parent::afterDelete();
 	}
